Use set_array if chunks already exist in store
When a predefined store is used, make sure to call add_chunk only when
there's more structures in has_structure then chunks already exist in
store.  Assuming that the given store already holds some information
about the structures in has_structure, this ensures that the neighbor
information lines up with that.