Return consistent indices from StructureStorage
When not specifying indices when creating an Atoms, the mapping between
chemical element and index is not unique and may change.  For
calculations on single structures this is no problem, but when iterating
over a StructureStorage in a single (e.g. interactive) calculation this
may cause confusion, since the atomistic jobs generally only save the
indices of atoms across multiple steps.