Raise error when incompatible structures are set on interactive jobs
If the species_symbols between two structures don't match (including
order) the atom indices written to the job output will be inconsistent.
While the current code takes care to send the correct remapped indices
to lammps the unmapped indices are written to HDF5, therefore on reading
the job output thing will no longer match.

Until this is fixed this error is raised to alert users, that the
results will not match what they expect.