Avoid calling get_chemical_elements in Atoms.select_index
select_index needs to check which Atoms in a structure match the passed
in element.  It does so by calling get_chemical_elements and then
compares the ChemicalElement object of each atom in the structure.
get_chemical_elements recreates an array of the chemical element objects
every time it is called.  That costs memory for large structures and is
unnecesary since Atoms.indices keeps a perfectly good array of chemical
indices around at all times.  Therefore the new code translates the
passed in element(s) into the index first then looks up which indices
match the ones in (the already existing) Atoms.indices.