Add a new structure factory for the MaterialsProject
Only two API calls are wrapped for now:
1. the summary search
2. get a structure by ID

More APIs to obtain structure exist (grain boundaries..), but should be
implemented as needed.

The search method returns a custom object implementing HasStructure.
This is done so that we can add methods here that convert the results to
other convenient objects, like:
- plain list
- StructureStructure
- pandas dataframe?