Use a separate ase structure copy in AseJob
Previously AseJob used the Atoms class from ASE internally.  This breaks
expectation of other code in pyiron that `job.structure` is always our
subclass of the Atoms class.  The previous commit added an automatic
conversion from ASE Atoms to ours, but that breaks the internal code of
this class.  As a work around this commit introduces a new attribute
`_ase_atoms` that keeps an ASE Atoms with the calculator attached and
changes the standard `job.structure` to be our Atoms sub class.  They
are kept in sync with the structure setter.