Be ASE compatible wrt spins
ASE calculator objects (in particular GPAW) have some expectations about
the spins.  Atoms should raise an AttributeError on accessing `.spins`
if none are defined and `get_initial_magnetic_moments` should returns
zeros.  If that is not the case then calculators may not reliably detect
that no spins are set and attempt calculations involving NaNs which
produce weird errors down the line.

With these two small changes the pyiron Atoms object can be used with
the GPAW calculator as if it was a normal ASE Atoms.  I assume any other
calculator should also be working.