Make Atoms pickle-able
Previous problems with {,un}pickling Atoms were caused by our unclean
overriding of `__getattr__`.  pickle (copy presumably as well) relies on
this method to discover where the object in question defines methods
that customize this behaviour  (`__reduce_ex__` and `__setstate__`).

That means first that returning `None` indiscriminately on undefined
keys from `__getattr__` will break this search, because pickle will try
to call `None`s.

Second `__setstate__` will be called on an object just
after `__new__` (on copied and unpickled object `__init__` is not called
at all).  So when `__getattr__` relies on attributes itself that are
only defined in `__init__` this will lead to an infinite regression.

Both of these problems appear on the `ChemicalElements` and
`PeriodicTable` classes as used inside `Atoms` and also with `Atoms`
itself.  For the first two I have simply removed `__getattr__` and
adjusted down-stream code because users are not (or rarely) interacting
with it and the additional headache is not worth it, just to cut down
some typing.  For `Atoms` accessing tags as attributes is part of the
interface, so I have moved those attributes (_tag_list) that cause the
regress to be initialized in `__new__`.