VASP: Parse the number of steps correctly from OUTCAR
For VASP jobs output/generic/steps is generated in either of two ways:
1. vasprun.xml can be parsed -> np.arange(len(output/generic/energy_tot))
2. vasprun.xml cannot be parsed -> np.linspace(0, #trigger)

where #trigger is the number of times the string

FREE ENERGIE OF THE ION-ELECTRON SYSTEM (eV)

occurs in the OUTCAR file.

The latter leads to fractional output and always 50 entries in that
array.  Both is contradictory to the first way and seems absurd.  This
commit changes this to align with the first way.  Also this fixes a bug
where the intent of the previous code seems to have been to multiply the
step indices by the value of NBLOCK.  This however is not done in the
first way.