Make generic energy_pot of Vasp configurable
Vasp outputs three energies, smeared internal energy, free energy and
smeared internal extrapolated to zero smearing.
This change allows to pick which of them is set as

output/generic/energy_pot

of the job.

Current defaults are smeared internal extrapolated to zero smearing when
vasprun.xml can be read and free energy when OUTCAR is read.